API fixes and folder restructure
