+++ src/main/java/org/project/map/controller/MapPointRestController.java
