+++ src/main/java/org/project/map/repository/MapPointRepository.java
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package org.project.uauth.user.dao.impl;
+package org.project;
 
 import org.project.map.model.MapPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
